move SingleValueAttribute to Manager
