+++ src/es6/canvas_widget/Manager.js
@@ -41,13 +41,17 @@ import { default as FileAttribute } from "./FileAttribute";
 import KeySelectionValueSelectionValueListAttribute from "./KeySelectionValueSelectionValueListAttribute";
 import QuizAttribute from "./QuizAttribute";
 import { default as SingleMultiLineValueAttribute } from "./SingleMultiLineValueAttribute";
-import { default as SingleValueAttribute } from "./SingleValueAttribute";
 import SingleValueListAttribute from "./SingleValueListAttribute";
 import ViewEdge from "./view/ViewEdge";
 import ViewNode from "./view/ViewNode";
 import LogicalConjunctions from "./viewpoint/LogicalConjunctions";
 import LogicalOperator from "./viewpoint/LogicalOperator";
 import ViewTypesUtil from "./viewpoint/ViewTypesUtil";
+
+const canvasSingleValueAttributeHtml = await loadHTML(
+  "../../templates/canvas_widget/single_value_attribute.html",
+  import.meta.url
+);
 
 const listHtml = await loadHTML(
   "../../templates/canvas_widget/list_attribute.html",
@@ -8160,6 +8164,95 @@ export class IntegerValue extends AbstractValue {
     init();
   }
 }
+
+/**
+ * SingleValueAttribute
+ * @class canvas_widget.SingleValueAttribute
+ * @extends canvas_widget.AbstractAttribute
+ * @memberof canvas_widget
+ * @constructor
+ * @param {string} id Entity id
+ * @param {string} name Name of attribute
+ * @param {canvas_widget.AbstractEntity} subjectEntity Entity the attribute is assigned to
+ */
+export class SingleValueAttribute extends AbstractAttribute {
+  value;
+  constructor(id, name, subjectEntity, y) {
+    y = y || window.y;
+    if (!y) {
+      throw new Error("y is undefined");
+    }
+    super(id, name, subjectEntity);
+    var that = this;
+
+    /***
+     * Value object of value
+     * @type {canvas_widget.Value}
+     * @private
+     */
+    var _value = new Value(id, name, this, this.getRootSubjectEntity(), y);
+    this.value = _value;
+
+    /**
+     * jQuery object of DOM node representing the node
+     * @type {jQuery}
+     * @private
+     */
+    var _$node = $(_.template(canvasSingleValueAttributeHtml)());
+
+    /**
+     * Set Value object of value
+     * @param {canvas_widget.Value} value
+     */
+    this.setValue = function (value) {
+      _value = value;
+      this.value = value;
+    };
+
+    /**
+     * Get Value object of value
+     * @returns {canvas_widget.Value}
+     */
+    this.getValue = function () {
+      return _value;
+    };
+
+    /**
+     * jQuery object of DOM node representing the attribute
+     * @type {jQuery}
+     * @private
+     */
+    this.get$node = function () {
+      return _$node;
+    };
+
+    /**
+     * Get JSON representation of the attribute
+     * @returns {Object}
+     */
+    this.toJSON = function () {
+      var json = AbstractAttribute.prototype.toJSON.call(this);
+      json.value = _value.toJSON();
+      return json;
+    };
+
+    /**
+     * Set attribute value by its JSON representation
+     * @param json
+     */
+    this.setValueFromJSON = function (json) {
+      _value.setValueFromJSON(json.value);
+    };
+
+    this.registerYType = function () {
+      _value.registerYType();
+    };
+
+    _$node.find(".name").text(this.getName());
+    _$node.find(".value").append(_value.get$node());
+  }
+}
+
 
 /**
  * Value
