+++ src/es6/canvas_widget/guidance_modeling/SetPropertyGuidance.js
@@ -1,7 +1,7 @@
 import "https://unpkg.com/jquery@3.6.0/dist/jquery.js";
 import "https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.13.2/jquery-ui.min.js";
 import _ from "lodash-es";
-import SingleValueAttribute from "../SingleValueAttribute";
+import { SingleValueAttribute } from "../Manager";
 import { SingleSelectionAttribute } from "../Manager";
 import { IntegerAttribute } from "../Manager";
 import BooleanAttribute from "../BooleanAttribute";
